+++ dfd_editor/include/dfd_editor/MultCanvasWindow.h
@@ -16,20 +16,21 @@ public:
   explicit MultCanvasWindow(std::string title);
   void DrawContents() override;
 
-  auto CreateNewCanvas(const std::shared_ptr<Dfd> &dfd) -> int;
+  auto CreateNewCanvas(const std::shared_ptr<Dfd> &dfd) -> int64_t;
 
-  void OpenCanvas(int canvas_id);
+  void OpenCanvas(int64_t canvas_id);
 
   void LoadDfd(const std::shared_ptr<Dfd> &dfd);
 
 private:
   std::vector<std::shared_ptr<Dfd>> dfds_;
   std::vector<std::shared_ptr<EditorCanvas>> canvas_;
-  auto AddCanvas(const std::shared_ptr<EditorCanvas> &node_editor) -> int;
+  auto AddCanvas(const std::shared_ptr<EditorCanvas> &node_editor) -> int64_t;
 
   bool has_connect_signal_ = false;
   void ConnectSignal();
-  int selected_canvas_id_ = 0; // 选择完之后会取负数，获取当前选择 tab 取绝对值
+  int64_t selected_canvas_id_ =
+      0; // 选择完之后会取负数，获取当前选择 tab 取绝对值
 
 public:
   MultCanvasWindow(MultCanvasWindow &&) = default;
