+++ dfd_editor/node_model/include/node_model/data/DataProcessData.h
@@ -6,31 +6,33 @@
 
 class DataProcessData {
 public:
-  DataProcessData(std::string *description, int canvas_id)
+  DataProcessData(std::string *description, int64_t canvas_id)
       : processing_description_(description),
         sub_data_flow_diagram_id_(canvas_id) {
   }
 
-  void AddInputDataFlow(int node_id) {
+  void AddInputDataFlow(uint64_t node_id) {
     input_data_flows_.push_back(node_id);
   }
-  void AddOutputDataFlow(int node_id) {
+  void AddOutputDataFlow(uint64_t node_id) {
     output_data_flows_.push_back(node_id);
   }
-  void RemoveInputDataFlow(int node_id);
-  void RemoveOutputDataFlow(int node_id);
+  void RemoveInputDataFlow(uint64_t node_id);
+  void RemoveOutputDataFlow(uint64_t node_id);
 
 private:
-  std::vector<int> input_data_flows_;
-  std::vector<int> output_data_flows_;
+  std::vector<uint64_t> input_data_flows_;
+  std::vector<uint64_t> output_data_flows_;
   std::string *processing_description_;
-  int sub_data_flow_diagram_id_;
+  int64_t sub_data_flow_diagram_id_;
 
 public:
-  [[nodiscard]] auto GetInputDataFlows() const -> const std::vector<int> & {
+  [[nodiscard]] auto GetInputDataFlows() const
+      -> const std::vector<uint64_t> & {
     return input_data_flows_;
   }
-  [[nodiscard]] auto GetOutputDataFlows() const -> const std::vector<int> & {
+  [[nodiscard]] auto GetOutputDataFlows() const
+      -> const std::vector<uint64_t> & {
     return output_data_flows_;
   }
 
@@ -44,10 +46,10 @@ public:
     *processing_description_ = content;
   }
 
-  void SetSubDataFlowDiagramId(int canvas_id) {
+  void SetSubDataFlowDiagramId(int64_t canvas_id) {
     sub_data_flow_diagram_id_ = canvas_id;
   }
-  [[nodiscard]] auto GetSubDataFlowDiagramId() const -> int {
+  [[nodiscard]] auto GetSubDataFlowDiagramId() const -> uint64_t {
     return sub_data_flow_diagram_id_;
   }
 };
