+++ dfd_editor/signal/include/signal/SignalHandel.h
@@ -20,7 +20,7 @@ public:
 
   sigslot::signal<int, const std::string &> processing_content_changed_;
   sigslot::signal<int> navigate_sub_diagram_onclick_;
-  sigslot::signal<const std::shared_ptr<Dfd> &, int &> create_new_dfd_;
+  sigslot::signal<const std::shared_ptr<Dfd> &, int64_t &> create_new_dfd_;
 
 private:
   SignalHandel() = default;
